requirements.txt
pandas 
notebook
numpy
scikit-learn
tensorflow
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
#-e .